Generalize slice widget to support 3D viewers. Move to viewers.common subpackage.
