+++ glue_qt/viewers/image/options_widget.py
@@ -4,7 +4,7 @@ from qtpy import QtWidgets
 
 from echo.qt import autoconnect_callbacks_to_qt
 from glue_qt.utils import load_ui, fix_tab_widget_fontsize
-from glue_qt.viewers.image.slice_widget import MultiSliceWidgetHelper
+from glue_qt.viewers.common.slice_widget import MultiSliceWidgetHelper
 from glue.viewers.matplotlib.state import MatplotlibDataViewerState
 
 __all__ = ['ImageOptionsWidget']
